backend/requirements.txt
Django>=3.1.12,<4.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0
mysqlclient>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.0
python-dateutil>=2.8.0
